Restructure under src folder
