+++ package.json
@@ -1,84 +1,84 @@
 {
-  "name": "react-typescript-fullstack-starter",
-  "version": "1.0.0",
-  "description": "A starter template for a Full-Stack application with react, typescript, webpack, and babel",
-  "main": "build/server/index.js",
-  "scripts": {
-    "start": "NODE_ENV=production node build/index.js",
-    "server": "npm run validate && nodemon build/index.js",
-    "client": "webpack-dev-server --open --config webpack.dev.js",
-    "dev": "npm run clean && concurrently \"npm run server\" \"npm run client\"",
-    "heroku-postbuild": "NPM_CONFIG_PRODUCTION=true npm run build",
-    "lint": "eslint --ignore-path .gitignore --ext .js,.ts,.tsx,.jsx .",
-    "prettier": "prettier --ignore-path .gitignore \"**/*.+(js|json|jsx|ts|tsx)\"",
-    "format": "npm run prettier -- --write",
-    "check-format": "npm run prettier -- --list-different",
-    "check-types": "tsc",
-    "validate": "npm-run-all --parallel check-types check-format lint build:server",
-    "build:client": "webpack --config webpack.prod.js",
-    "build:server": "babel server --extensions .js,.ts,.tsx --out-dir build",
-    "build": "npm run build:server && npm run build:client",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "clean": "rimraf build"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/frendon001/react-typescript-fullstack-starter.git"
-  },
-  "author": "Fausto Rendon",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/frendon001/react-typescript-fullstack-starter/issues"
-  },
-  "homepage": "https://github.com/frendon001/react-typescript-fullstack-starter#readme",
-  "dependencies": {
-    "@types/express": "^4.17.6",
-    "@types/helmet": "0.0.47",
-    "@types/react-router-dom": "^5.1.5",
-    "axios": "^0.19.2",
-    "body-parser": "^1.19.0",
-    "concurrently": "^5.2.0",
-    "express": "^4.17.1",
-    "helmet": "^3.22.0",
-    "nodemon": "^2.0.4",
-    "path-parser": "^6.1.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-hot-loader": "^4.12.21",
-    "react-router-dom": "^5.2.0"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.8.4",
-    "@babel/core": "^7.9.6",
-    "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/preset-env": "^7.9.6",
-    "@babel/preset-react": "^7.9.4",
-    "@babel/preset-typescript": "^7.9.0",
-    "@types/react": "^16.9.35",
-    "@types/react-dom": "^16.9.8",
-    "@types/webpack-env": "^1.15.2",
-    "@typescript-eslint/eslint-plugin": "^3.0.0",
-    "@typescript-eslint/parser": "^3.0.0",
-    "babel-loader": "^8.1.0",
-    "css-loader": "^3.5.3",
-    "eslint": "^7.1.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-react": "^7.20.0",
-    "file-loader": "^6.0.0",
-    "html-webpack-plugin": "^4.3.0",
-    "html-webpack-template": "^6.2.0",
-    "husky": "^4.2.5",
-    "lint-staged": "^10.2.6",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.0.5",
-    "rimraf": "^3.0.2",
-    "style-loader": "^1.2.1",
-    "terser-webpack-plugin": "^3.0.1",
-    "typescript": "3.7.5",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^4.2.2"
-  }
+	"name": "react-typescript-fullstack-starter",
+	"version": "1.0.0",
+	"description": "A starter template for a Full-Stack application with react, typescript, webpack, and babel",
+	"main": "build/server/index.js",
+	"scripts": {
+		"start": "NODE_ENV=production node build/index.js",
+		"server": "npm run validate && nodemon build/index.js",
+		"client": "webpack-dev-server --open --config webpack.dev.js",
+		"dev": "npm run clean && concurrently \"npm run server\" \"npm run client\"",
+		"heroku-postbuild": "NPM_CONFIG_PRODUCTION=true npm run build",
+		"lint": "eslint --ignore-path .gitignore --ext .js,.ts,.tsx,.jsx .",
+		"prettier": "prettier --ignore-path .gitignore \"**/*.+(js|json|jsx|ts|tsx)\"",
+		"format": "npm run prettier -- --write",
+		"check-format": "npm run prettier -- --list-different",
+		"check-types": "tsc",
+		"validate": "npm-run-all --parallel check-types check-format lint build:server",
+		"build:client": "webpack --config webpack.prod.js",
+		"build:server": "babel src --extensions .js,.ts,.tsx --out-dir build",
+		"build": "npm run build:server && npm run build:client",
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"clean": "rimraf build"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/frendon001/react-typescript-fullstack-starter.git"
+	},
+	"author": "Fausto Rendon",
+	"license": "MIT",
+	"bugs": {
+		"url": "https://github.com/frendon001/react-typescript-fullstack-starter/issues"
+	},
+	"homepage": "https://github.com/frendon001/react-typescript-fullstack-starter#readme",
+	"dependencies": {
+		"@types/express": "^4.17.6",
+		"@types/helmet": "0.0.47",
+		"@types/react-router-dom": "^5.1.5",
+		"axios": "^0.19.2",
+		"body-parser": "^1.19.0",
+		"concurrently": "^5.2.0",
+		"express": "^4.17.1",
+		"helmet": "^3.22.0",
+		"nodemon": "^2.0.4",
+		"path-parser": "^6.1.0",
+		"react": "^16.13.1",
+		"react-dom": "^16.13.1",
+		"react-hot-loader": "^4.12.21",
+		"react-router-dom": "^5.2.0"
+	},
+	"devDependencies": {
+		"@babel/cli": "^7.8.4",
+		"@babel/core": "^7.9.6",
+		"@babel/plugin-proposal-class-properties": "^7.8.3",
+		"@babel/preset-env": "^7.9.6",
+		"@babel/preset-react": "^7.9.4",
+		"@babel/preset-typescript": "^7.9.0",
+		"@types/react": "^16.9.35",
+		"@types/react-dom": "^16.9.8",
+		"@types/webpack-env": "^1.15.2",
+		"@typescript-eslint/eslint-plugin": "^3.0.0",
+		"@typescript-eslint/parser": "^3.0.0",
+		"babel-loader": "^8.1.0",
+		"css-loader": "^3.5.3",
+		"eslint": "^7.1.0",
+		"eslint-config-prettier": "^6.11.0",
+		"eslint-plugin-prettier": "^3.1.3",
+		"eslint-plugin-react": "^7.20.0",
+		"file-loader": "^6.0.0",
+		"html-webpack-plugin": "^4.3.0",
+		"html-webpack-template": "^6.2.0",
+		"husky": "^4.2.5",
+		"lint-staged": "^10.2.6",
+		"npm-run-all": "^4.1.5",
+		"prettier": "^2.0.5",
+		"rimraf": "^3.0.2",
+		"style-loader": "^1.2.1",
+		"terser-webpack-plugin": "^3.0.1",
+		"typescript": "3.7.5",
+		"webpack": "^4.43.0",
+		"webpack-cli": "^3.3.11",
+		"webpack-dev-server": "^3.11.0",
+		"webpack-merge": "^4.2.2"
+	}
 }
