+++ src/server/api/dummyApi/index.ts
@@ -4,7 +4,7 @@ const router = express.Router();
 
 // server for dummy API route
 router.get('/test', (req, res: express.Response) => {
-  res.send('Success');
+	res.send('Success');
 });
 
 export default router;
