+++ src/server/index.ts
@@ -10,23 +10,23 @@ app.use(helmet());
 app.use(bodyParser.json());
 // Allow cross-domain requests
 app.use((_req, res, next) => {
-  res.header('Access-Control-Allow-Origin', '*');
-  res.header(
-    'Access-Control-Allow-Headers',
-    'Origin, X-Requested-With, Content-Type, Accept',
-  );
-  next();
+	res.header('Access-Control-Allow-Origin', '*');
+	res.header(
+		'Access-Control-Allow-Headers',
+		'Origin, X-Requested-With, Content-Type, Accept',
+	);
+	next();
 });
 
 app.use('/api', api);
 
 if (process.env.NODE_ENV === 'production') {
-  // Serve production assets
-  app.use(express.static('build'));
-  console.log(__dirname);
-  app.get('/*', (_req, res) => {
-    res.sendFile(path.resolve(__dirname, 'generated.html'));
-  });
+	// Serve production assets
+	app.use(express.static('build'));
+	console.log(__dirname);
+	app.get('/*', (_req, res) => {
+		res.sendFile(path.resolve(__dirname, 'generated.html'));
+	});
 }
 
 const PORT = process.env.PORT || 3030;
