+++ webpack.prod.js
@@ -4,33 +4,33 @@ const TerserPlugin = require('terser-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = merge(common, {
-  mode: 'production',
-  plugins: [
-    new HtmlWebpackPlugin({
-      filename: 'generated.html',
-      title: '[APP NAME]',
-      meta: [
-        // viewport meta tag for responsiveness / media queries
-        {
-          name: 'viewport',
-          content: 'width=device-width, initial-scale=1',
-        },
-      ],
-      template: require('html-webpack-template'),
-      favicon: './app/public/favicon.ico',
-      appMountId: 'root',
-      minify: {
-        removeComments: true,
-        collapseWhitespace: true,
-        conservativeCollapse: true,
-      },
-    }),
-  ],
-  optimization: {
-    splitChunks: {
-      chunks: 'initial',
-    },
-    minimize: true,
-    minimizer: [new TerserPlugin()],
-  },
+	mode: 'production',
+	plugins: [
+		new HtmlWebpackPlugin({
+			filename: 'generated.html',
+			title: '[APP NAME]',
+			meta: [
+				// viewport meta tag for responsiveness / media queries
+				{
+					name: 'viewport',
+					content: 'width=device-width, initial-scale=1',
+				},
+			],
+			template: require('html-webpack-template'),
+			favicon: './app/public/favicon.ico',
+			appMountId: 'root',
+			minify: {
+				removeComments: true,
+				collapseWhitespace: true,
+				conservativeCollapse: true,
+			},
+		}),
+	],
+	optimization: {
+		splitChunks: {
+			chunks: 'initial',
+		},
+		minimize: true,
+		minimizer: [new TerserPlugin()],
+	},
 });
